implements 2 parries in case that ball comes straight towards the keeper:
-parry_straight: pretty safe
-parry_straight1: more beautiful, but need 100% confidence that ball is coming extremely straight.

i would suggest to take always the parry_straight.

Please notice, that for now the nao will wait 4sec in the parry-pose to keep the ball and stand up again afterwards.
feel free to ask to change that delay. It's still possible to make those delays nonexisting, and only perform standup when perception allows it.
BOTH parrys I just implemented should have enough vision to work with the perception.

Sorry for the wall of text ;)